+++ client/my-sites/plan-features-2023-grid/style.scss
@@ -1044,27 +1044,6 @@ body.is-section-signup.is-white-signup,
 	}
 }
 
-/**
- * We auto-scroll to the plan comparison grid when it is shown.
- * However, on the /plans page, due to the presence of the masterbar on top,
- * we need to add a scroll offset.
- */
-.is-section-plans .plan-features-2023-grid__plan-comparison-grid-container {
-	scroll-margin-top: $plan-features-header-banner-height + 24px;
-	margin: auto;
-	@include plan-comparison-grid-breakpoints-with-sidebar;
-}
-
-.is-section-signup .plan-features-2023-grid__plan-comparison-grid-container {
-	margin: auto;
-	@include plan-comparison-grid-breakpoints;
-}
-
-.is-section-stepper .plan-features-2023-grid__plan-comparison-grid-container {
-	margin: auto;
-	@include plan-comparison-grid-breakpoints;
-}
-
 .formatted-header {
 	margin-bottom: 0;
 }
